+++ crates/bevy_auto_plugin_shared/src/syntax/extensions/mod.rs
@@ -1,2 +1,3 @@
 pub mod item;
+pub mod lit;
 pub mod path;