+++ CHANGELOG.md
@@ -103,6 +103,8 @@
 
 ---
 ### v0.7.0
+- Add `name` param to `auto_name` attribute
+- Add the ability to set a custom name with `auto_name` param in `auto_component`
 - Expanded `auto_insert_resource` `resource` to accept more potentially valid expressions
 - Remove re-exported crates `bevy_ecs`, `bevy_ecs_macros`, `bevy_state`, `bevy_log`, `bevy_reflect`, `bevy_reflect_derive`
   - Fixes